Extract Membership into it's own class
